refactor(persistence): nest remaining root types into folders by sub-namespace (#905)
Finishes nesting the partially-organised Persistence class library (implementation)
so it is no longer top-heavy, continuing the folder-by-sub-namespace pattern from
#896-#904.

Moves the 15 remaining root-level types (71% of 21 files) into technical folders
matching house style (folder = sub-namespace), mirroring Persistence.Abstractions
(#902), and leaving the existing Lookups/Extensions folders in this assembly
untouched:

- Configuration/ - CosmosDbSettings
- Factories/ - CosmosDbClientFactory, CosmosDbContainerFactory, FileRepositoryFactory
- Providers/ - JsonSerializerOptionsProvider
- Repositories/ - PodcastRepository, LookupRepository, FileRepository,
  EpisodeRepository, EliminationTermsRepository, ActivityRepository
- Serialization/ - CosmosSystemTextJsonSerializer, RegexConverter
- Writers/ - SafeFileEntityWriter (+ interface)

Root .cs count: 15 -> 0 (21 files total, unchanged). No leftovers at the assembly
root.

Public type names and the AddRepositories() / AddEliminationTerms() /
AddFileRepository() / AddSafeFileWriter() DI extension methods are unchanged.
Updated consumer usings across Console-Apps (FindDuplicateEpisodes,
CultPodcasts.DatabasePublisher, CosmosDbDownloader, CreateSearchIndex) - new
explicit using directives were needed because C#'s enclosing-namespace fallback
lookup no longer reaches types once they move out of the assembly root namespace
into new sibling sub-namespaces. Also removed 5 stale/unused
`using RedditPodcastPoster.Persistence;` directives in
RedditPodcastPoster.Episodes.Tests that predated this change. No namespace
collisions with Persistence.Abstractions.Repositories/Providers/Factories or
RedditPodcastPoster.Models.

Base includes #904 (refactor(subjects)); branched from updated main. Did not
touch ContentPublisher, Models, or Durable hosts.

Test plan:
- RedditPodcastPoster.Persistence.Tests: 7/7 passed
- RedditPodcastPoster.Episodes.Tests: 291/291 passed
- RedditPodcastPoster.PodcastServices.Tests: 57/57 passed
- FunctionHost.Tests: 24/24 passed
- Full solution build (RedditPodcastPoster.slnx): green, 0 errors
- No linter errors on touched files
