+++ Class-Libraries/RedditPodcastPoster.Episodes.Tests/BusinessRules/Matching/FuzzyTitleMatchingRules.cs
@@ -4,8 +4,6 @@ using RedditPodcastPoster.Episodes.TestSupport;
 using RedditPodcastPoster.Episodes.TestSupport.Assertions;
 using RedditPodcastPoster.Episodes.TestSupport.Fixtures;
 using RedditPodcastPoster.Models;
-using RedditPodcastPoster.Persistence;
-
 namespace RedditPodcastPoster.Episodes.Tests.BusinessRules.Matching;
 
 /// <summary>
