+++ Class-Libraries/RedditPodcastPoster.Episodes.Tests/BusinessRules/Merging/ReleaseDateMergingRules.cs
@@ -3,7 +3,6 @@ using RedditPodcastPoster.Episodes.TestSupport;
 using RedditPodcastPoster.Episodes.TestSupport.Assertions;
 using RedditPodcastPoster.Episodes.TestSupport.Fixtures;
 using RedditPodcastPoster.Models;
-using RedditPodcastPoster.Persistence;
 
 namespace RedditPodcastPoster.Episodes.Tests.BusinessRules.Merging;
