+++ Class-Libraries/RedditPodcastPoster.Persistence/Configuration/CosmosDbSettings.cs
@@ -1,4 +1,4 @@
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Configuration;
 
 public class CosmosDbSettings
 {
