+++ Class-Libraries/RedditPodcastPoster.Persistence/Extensions/ServiceCollectionExtensions.cs
@@ -5,7 +5,12 @@ using RedditPodcastPoster.DependencyInjection;
 using RedditPodcastPoster.Persistence.Abstractions.Factories;
 using RedditPodcastPoster.Persistence.Abstractions.Providers;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
+using RedditPodcastPoster.Persistence.Configuration;
+using RedditPodcastPoster.Persistence.Factories;
 using RedditPodcastPoster.Persistence.Lookups;
+using RedditPodcastPoster.Persistence.Providers;
+using RedditPodcastPoster.Persistence.Repositories;
+using RedditPodcastPoster.Persistence.Writers;
 using RedditPodcastPoster.PodcastServices.Abstractions;
 using RedditPodcastPoster.Text.EliminationTerms;
 using RedditPodcastPoster.Text.KnownTerms;
