+++ Class-Libraries/RedditPodcastPoster.Persistence/Providers/JsonSerializerOptionsProvider.cs
@@ -1,8 +1,9 @@
 ﻿using System.Text.Json;
 using System.Text.Json.Serialization;
 using RedditPodcastPoster.Persistence.Abstractions.Providers;
+using RedditPodcastPoster.Persistence.Serialization;
 
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Providers;
 
 public class JsonSerializerOptionsProvider : IJsonSerializerOptionsProvider
 {
