+++ Class-Libraries/RedditPodcastPoster.Persistence/Repositories/EliminationTermsRepository.cs
@@ -2,7 +2,7 @@
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
 
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Repositories;
 
 public class EliminationTermsRepository(
     ILookupRepository lookupRepository,
