+++ Class-Libraries/RedditPodcastPoster.Persistence/Repositories/EpisodeRepository.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Logging;
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
 
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Repositories;
 
 public class EpisodeRepository(
     Container container,
