+++ Class-Libraries/RedditPodcastPoster.Persistence/Repositories/LookupRepository.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Logging;
 using RedditPodcastPoster.Models;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
 
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Repositories;
 
 public class LookupRepository(
     Container lookupContainer,
