+++ Class-Libraries/RedditPodcastPoster.Persistence/Serialization/RegexConverter.cs
@@ -2,7 +2,7 @@
 using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
 
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Serialization;
 
 public class RegexConverter : JsonConverter<Regex>
 {
