+++ Class-Libraries/RedditPodcastPoster.Persistence/Writers/ISafeFileEntityWriter.cs
@@ -1,6 +1,6 @@
 ﻿using RedditPodcastPoster.Models;
 
-namespace RedditPodcastPoster.Persistence;
+namespace RedditPodcastPoster.Persistence.Writers;
 
 public interface ISafeFileEntityWriter
 {
