+++ Console-Apps/FindDuplicateEpisodes/FindDuplicateEpisodesProcessor.cs
@@ -4,9 +4,9 @@ using Microsoft.Azure.Cosmos;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using RedditPodcastPoster.Models;
-using RedditPodcastPoster.Persistence;
 using RedditPodcastPoster.Persistence.Abstractions.Providers;
 using RedditPodcastPoster.Persistence.Abstractions.Repositories;
+using RedditPodcastPoster.Persistence.Configuration;
 
 namespace FindDuplicateEpisodes;
